refactor: move DataFormat enum to cli module
