+++ synth/src/cli/mod.rs
@@ -301,6 +301,42 @@ pub enum TelemetryCommand {
     Status,
 }
 
+#[derive(Debug, Clone)]
+pub enum DataFormat {
+    Json,
+    JsonLines {
+        collection_field_name: Option<String>,
+    },
+    Csv,
+}
+
+impl DataFormat {
+    pub fn new(uri_scheme: &str, collection_field_name: Option<String>) -> Self {
+        match uri_scheme {
+            "jsonl" => DataFormat::JsonLines {
+                collection_field_name,
+            },
+            "csv" => DataFormat::Csv,
+            _ => DataFormat::Json,
+        }
+    }
+
+    pub fn get_collection_field_name_or_default(&self) -> &str {
+        match self {
+            DataFormat::JsonLines {
+                collection_field_name: Some(ref x),
+            } => x,
+            _ => "collection", // Default collection field name is 'collection'.
+        }
+    }
+}
+
+impl Default for DataFormat {
+    fn default() -> Self {
+        DataFormat::Json
+    }
+}
+
 #[cfg(test)]
 pub mod tests {
     use super::*;
